+++ bmc-objectstorage/bmc-objectstorage-generated/src/main/java/com/oracle/bmc/objectstorage/internal/http/CreateBucketConverter.java
@@ -98,21 +98,21 @@ public class CreateBucketConverter {
 
                                 com.google.common.base.Optional<java.util.List<String>> eTagHeader =
                                         com.oracle.bmc.http.internal.HeaderUtils.get(
-                                                headers, "eTag");
+                                                headers, "ETag");
                                 if (eTagHeader.isPresent()) {
                                     builder.eTag(
                                             com.oracle.bmc.http.internal.HeaderUtils.toValue(
-                                                    "eTag", eTagHeader.get().get(0), String.class));
+                                                    "ETag", eTagHeader.get().get(0), String.class));
                                 }
 
                                 com.google.common.base.Optional<java.util.List<String>>
                                         locationHeader =
                                                 com.oracle.bmc.http.internal.HeaderUtils.get(
-                                                        headers, "location");
+                                                        headers, "Location");
                                 if (locationHeader.isPresent()) {
                                     builder.location(
                                             com.oracle.bmc.http.internal.HeaderUtils.toValue(
-                                                    "location",
+                                                    "Location",
                                                     locationHeader.get().get(0),
                                                     String.class));
                                 }