+++ bmc-objectstorage/bmc-objectstorage-generated/src/main/java/com/oracle/bmc/objectstorage/internal/http/GetObjectLifecyclePolicyConverter.java
@@ -109,11 +109,11 @@ public class GetObjectLifecyclePolicyConverter {
 
                                 com.google.common.base.Optional<java.util.List<String>> eTagHeader =
                                         com.oracle.bmc.http.internal.HeaderUtils.get(
-                                                headers, "eTag");
+                                                headers, "ETag");
                                 if (eTagHeader.isPresent()) {
                                     builder.eTag(
                                             com.oracle.bmc.http.internal.HeaderUtils.toValue(
-                                                    "eTag", eTagHeader.get().get(0), String.class));
+                                                    "ETag", eTagHeader.get().get(0), String.class));
                                 }
 
                                 GetObjectLifecyclePolicyResponse responseWrapper = builder.build();